fix: BE resolve 'this' context problem in user schema
This commit resolves the problem with the 'matchPassword' method in the
user schema, which was originally declared using an arrow function,
causing issues with the 'this' keyword during user logins. It also
corrects a minor typo from 'matchPassWord' to 'matchPassword' when
invoking the method inside the authUser function in userController.js.